+++ FirebaseStorage/Tests/Unit/StorageAPITests.swift
@@ -67,18 +67,6 @@ final class StorageAPITests: XCTestCase {
     _ = ref.putFile(from: file, metadata: metadata) { (result: StorageMetadata?, error: Error?) in
     }
 
-    let fileHandle = FileHandle.nullDevice
-    _ = ref.putFileHandle(fileHandle)
-    _ = ref.putFileHandle(fileHandle, metadata: metadata)
-    _ = ref.putFileHandle(fileHandle, metadata: metadata) { (result: Result<
-      StorageMetadata,
-      Error
-    >) in
-    }
-    _ = ref.putFileHandle(fileHandle, metadata: metadata) { (result: StorageMetadata?,
-                                                             error: Error?) in
-    }
-
     _ = ref.getData(maxSize: 122) { (result: Data?, error: Error?) in
     }
     _ = ref.getData(maxSize: 122) { (result: Result<Data, Error>) in